Fix implicit conversion of returned value in AssetLibrary::numMaterials()

+++ framework/include/minko/file/AssetLibrary.hpp
@@ -99,7 +99,7 @@ namespace minko
             uint
             numMaterials()
             {
-                return _materials.size();
+                return static_cast<uint>(_materials.size());
             }
 
             inline
